# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Server Owners Support API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  discordId: string;
  username: string;
  /** @nullable */
  globalName?: string | null;
  /** @nullable */
  avatar?: string | null;
  isAdmin: boolean;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  owner: boolean;
}

export interface Template {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  templateCode: string;
  category: string;
  createdBy: string;
  createdAt: string;
  featured: boolean;
}

export interface TemplateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @minLength 1 */
  templateCode: string;
  category: string;
  featured?: boolean;
}

export interface TemplateUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  templateCode?: string;
  category?: string;
  featured?: boolean;
}

export interface Admin {
  discordId: string;
  /** @nullable */
  addedBy?: string | null;
  createdAt: string;
}

export interface AdminInput {
  /** @minLength 1 */
  discordId: string;
}

export interface AdminStats {
  totalTemplates: number;
  totalAdmins: number;
  totalUsers: number;
  featuredTemplates: number;
}

export type ListTemplatesParams = {
category?: string;
featured?: boolean;
};