# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Server Owners Support API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: templates
    description: Discord templates
  - name: admin
    description: Admin management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /discord/guilds:
    get:
      operationId: getMyGuilds
      tags: [auth]
      summary: Get user's Discord guilds where they are owner
      responses:
        "200":
          description: List of guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new template (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: listAdmins
      tags: [admin]
      summary: List all admin users
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addAdmin
      tags: [admin]
      summary: Add a new admin by Discord ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Admin added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{discordId}:
    delete:
      operationId: removeAdmin
      tags: [admin]
      summary: Remove an admin
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    User:
      type: object
      properties:
        discordId:
          type: string
        username:
          type: string
        globalName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        isAdmin:
          type: boolean
      required:
        - discordId
        - username
        - isAdmin

    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        owner:
          type: boolean
      required:
        - id
        - name
        - owner

    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        templateCode:
          type: string
        category:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
        featured:
          type: boolean
      required:
        - id
        - name
        - description
        - templateCode
        - category
        - createdBy
        - createdAt
        - featured

    TemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        templateCode:
          type: string
          minLength: 1
        category:
          type: string
        featured:
          type: boolean
      required:
        - name
        - description
        - templateCode
        - category

    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        templateCode:
          type: string
        category:
          type: string
        featured:
          type: boolean

    Admin:
      type: object
      properties:
        discordId:
          type: string
        addedBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - discordId
        - createdAt

    AdminInput:
      type: object
      properties:
        discordId:
          type: string
          minLength: 1
      required:
        - discordId

    AdminStats:
      type: object
      properties:
        totalTemplates:
          type: integer
        totalAdmins:
          type: integer
        totalUsers:
          type: integer
        featuredTemplates:
          type: integer
      required:
        - totalTemplates
        - totalAdmins
        - totalUsers
        - featuredTemplates