# lib/api-zod/src/generated/types/guild.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Server Owners Support API
 * OpenAPI spec version: 0.1.0
 */

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  owner: boolean;
}