# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Server Owners Support API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @minLength 1 */
  templateCode: string;
  category: string;
  featured?: boolean;
}