# lib/api-zod/src/generated/types/templateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Server Owners Support API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  templateCode?: string;
  category?: string;
  featured?: boolean;
}