# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Server Owners Support API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  discordId: string;
  username: string;
  /** @nullable */
  globalName?: string | null;
  /** @nullable */
  avatar?: string | null;
  isAdmin: boolean;
}